requirements.txt
SQLAlchemy==1.4.37
pymysql